:wrench:  Pr follow-up rename DateInput component
